pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kvpress-eval"
version = "0.1.0"
description = "Workload-aware evaluation of KV-cache optimization methods with KVPress."
requires-python = ">=3.10"
dependencies = [
  "kvpress",
  "optimum-quanto",
  "pandas",
  "matplotlib",
  "PyYAML",
  "nltk",
  "bert-score",
  "rouge",
  "jieba",
  "fuzzywuzzy",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
